requirements.txt
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.2.0
prettytable==3.3.0
scikit_learn==1.1.2
seaborn==0.11.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
